Add option to load internal IAutoMappingOverrides
