+++ global.json
@@ -1,6 +1,6 @@
 {
     "projects": [ "src" ],
     "sdk": {
-        "version": "2.1.700"
+        "version": "2.1.812"
     }
 }
