+++ src/FluentNHibernate.Testing/Fixtures/AutoMappingAlterations/InternalOverride.cs
@@ -1,0 +1,14 @@
+﻿using FluentNHibernate.Automapping;
+using FluentNHibernate.Automapping.Alterations;
+using FluentNHibernate.Testing.Fixtures.AutoMappingAlterations.Model;
+
+namespace FluentNHibernate.Testing.Fixtures.AutoMappingAlterations
+{
+    internal class InternalOverride : IAutoMappingOverride<Baz>
+    {
+        public void Override(AutoMapping<Baz> mapping)
+        {
+            mapping.Table("InternalBaz");
+        }
+    }
+}
